Wrap lesson plans component in suspense
